requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
plotly
joblib
jupyterlab
xgboost
lightgbm
catboost
imbalanced-learn
mlflow
pymongo
python-dotenv
PyYAML
Flask
-e .